requirements.txt
fastapi
uvicorn[standard]
pandas
openpyxl
matplotlib
seaborn
python-multipart
python-pptx
reportlab
scikit-learn
google-generativeai
python-dotenv
